Refactor MP4 to text conversion: replace old script with new implementation and update .gitignore
