+++ .gitignore
@@ -28,3 +28,6 @@ venv.bak/
 
 #download 
 nas-download/
+
+#mp4 to txt
+mp4-to-txt/
